requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_socketio
python-socketio
eventlet
pyjwt
werkzeug